git_info: Add get_latest_commit function #57
Implement in all scripts
